+ stations have a two way space for trains + subwayspaces can have a list of trains passing + line creation creation made easier + bounce points for terminals + changed scheduler registrator contract to receive a string with the description + minor bug currently ends with exception in most of the cases, gonna debug it

+++ src/SimulatorObject.java
@@ -1,6 +1,5 @@
 
 public interface SimulatorObject {
 
-	
-	public void event(double timestamp) throws Exception;
+	public void event(Long timestamp) throws Exception;
 }
